+++ src/main/java/edu/oregonstate/cs361/battleship/BattleshipModel.java
@@ -1,7 +1,7 @@
 package edu.oregonstate.cs361.battleship;
 
-import java.util.ArrayList;
 import java.util.Random;
+import java.util.ArrayList;
 
 /**
  * Created by michaelhilton on 1/4/17.
@@ -25,8 +25,8 @@ public class BattleshipModel {
     /*
     Other variables the battleship model uses.
      */
-    ArrayList<Coordinate> playerHits;
-    ArrayList<Coordinate> playerMisses;
+    static ArrayList<Coordinate> playerHits;
+    static ArrayList<Coordinate> playerMisses;
     ArrayList<Coordinate> computerHits;
     ArrayList<Coordinate> computerMisses;
     ArrayList<Coordinate> playerSunk;
@@ -35,6 +35,132 @@ public class BattleshipModel {
 
     boolean scanResult = false;
     boolean hardMode = false;
+    public class Mode {
+        /*
+            This is a parent class for easy mode and hard mode. These empty functions will be overridden
+            by a function with the same name in either the easy or hard mode class.
+         */
+        public void placeShips(Ship computer_aircraftCarrier, Ship computer_battleship, Ship computer_submarine,Ship computer_Dinghy,Ship computer_Clipper){
+
+
+        }
+        public void fire(){
+
+        }
+
+    }
+    public class Hard extends Mode {
+
+        boolean hit;
+        int count;//tracks directions checked around a hit
+        Coordinate lastHit;
+        Coordinate hitPrime;
+
+
+        public Hard(){
+
+        }
+        private void _fire(Coordinate shot){//helper function for fire
+            int result;
+            result = playerShot(shot);//shoots at the player
+            if (result == 1) {//ship was hit but not sunk
+                if(!hit) {//ship not found already
+                    hitPrime.setAcross(shot.getAcross());//record location of ship
+                    hitPrime.setDown(shot.getDown());
+                }
+                hit = true;//note that a ship has been found
+                count = 0;//initialize counter
+                lastHit.setAcross(shot.getAcross());//record location of hit
+                lastHit.setDown(shot.getDown());
+            } else if (result == 2) {//ship was sunk
+                hit = false;//note that we have no information about a ship's location
+                count = 0;//reinitialize count
+            } else {//miss
+                if (hit)
+                    count = count + 1;//note failed direction if ship found
+                else
+                    count = 0;//do nothing
+            }
+        }
+        public void fire() {
+            int max = 10;//this section mostly copied from model's fire function
+            int min = 1;
+            Coordinate shot = new Coordinate(1, 1);
+            if (!hit) {//ship not found
+                do {
+                    Random random = new Random();
+                    int randRow = random.nextInt(max - min + 1) + min;
+                    int randCol = random.nextInt(max - min + 1) + min;
+                    shot.setAcross(randRow);
+                    shot.setDown(randCol);
+                } while (playerMisses.contains(shot) || (playerHits.contains(shot)));//don't fire on the same spot more than once
+                _fire(shot);
+            }
+            else {//ship found
+                int row = 0,col = 0;
+                do {
+                    switch (count) {//choose direction to move by count
+                        case 0:
+                            col = lastHit.getDown() + 1;
+                            row = lastHit.getAcross();
+                            break;
+                        case 1:
+                            col = lastHit.getDown();
+                            row = lastHit.getAcross() + 1;
+                            break;
+                        case 2:
+                            col = lastHit.getDown() - 1;
+                            row = lastHit.getAcross();
+                            break;
+                        case 3:
+                            col = lastHit.getDown();
+                            row = lastHit.getAcross() - 1;
+                            break;
+                        case 4://all four directions missed or already hit
+                            lastHit.setAcross(hitPrime.getAcross());//revert to original location
+                            lastHit.setDown(hitPrime.getDown());
+                            col = lastHit.getDown();//set shot to previous hit
+                            row = lastHit.getAcross();
+                            count = -1;//count will be incremented to zero by if
+                            break;
+                    }
+                    shot.setDown(col);
+                    shot.setAcross(row);
+                    if((playerMisses.contains(shot)) || (playerHits.contains(shot)))//inc count if shot previously taken
+                        count = count + 1;
+                } while ((playerMisses.contains(shot)) || (playerHits.contains(shot)));
+                _fire(shot);//FIRE
+            }
+        }
+
+
+    }
+
+    public class Easy extends Mode {
+
+        public Easy(){
+            //nothing to construct
+        }
+        /*
+            Easy mode hard codes the placement of the computer's ships.
+         */
+        public void placeShips(Military computer_aircraftCarrier,  Military computer_battleship,  Military computer_submarine, Civilian computer_Dinghy,  Civilian computer_Clipper){
+
+            computer_aircraftCarrier.setLocation(new Coordinate(1, 0),new Coordinate(1,4));
+            computer_battleship.setLocation(new Coordinate(2,0), new Coordinate(2, 3));
+            computer_Clipper.setLocation(new Coordinate(5,3),new Coordinate(5,3));
+            computer_Dinghy.setLocation(new Coordinate(3,7), new Coordinate(5, 7));
+            computer_submarine.setLocation(new Coordinate(7, 9), new Coordinate(9, 9));
+
+
+
+        }
+
+        public void fire(){
+
+        }
+
+    }
 
     private Mode gameMode;
