+++ src/main/java/edu/oregonstate/cs361/battleship/Mode.java
@@ -1,21 +1,10 @@
 package edu.oregonstate.cs361.battleship;
 
+import java.util.ArrayList;
+
 /**
  * Created by Keana on 3/15/2017.
  */
-public class Mode {
-    /*
-        This is a parent class for easy mode and hard mode. These empty functions will be overridden
-        by a function with the same name in either the easy or hard mode class.
-     */
-    public void placeShips(Military computer_aircraftCarrier,  Military computer_battleship,  Military computer_submarine, Civilian computer_Dinghy,  Civilian computer_Clipper){
-
-    }
-
-    public void fire(){
-
-    }
-
-}
+import java.util.concurrent.Callable;
